--- src/main/java/team/projectzebra/persistence/Repository/PlaceRepository.java
@@ -1,7 +1,0 @@
-package team.projectzebra.persistence.Repository;
-
-import org.springframework.data.repository.CrudRepository;
-import team.projectzebra.persistence.entity.Place;
-
-public interface PlaceRepository extends CrudRepository<Place,Integer> {
-}
