+++ src/main/java/team/projectzebra/persistence/repository/WorkplaceRepository.java
@@ -1,0 +1,7 @@
+package team.projectzebra.persistence.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import team.projectzebra.persistence.entity.Workplace;
+
+public interface WorkplaceRepository extends CrudRepository<Workplace,Integer> {
+}
